+++ SystemPaulinaVianaMachado/SystemPaulinaVianaMachado/src/main/java/com/grupoBlueSpectre/SystemPaulinaVianaMachado/Core/Usecases/Intefaces/DeleteUsecase.java
@@ -1,5 +1,5 @@
 package com.grupoBlueSpectre.SystemPaulinaVianaMachado.Core.Usecases.Intefaces;
 
 public interface DeleteUsecase {
-    void execute(int ID);
+    void execute(Long ID);
 }
